Move mapItemsToItemTrees() into item util
