+++ src/components/Menu/utils/item.js
@@ -40,3 +40,31 @@ export function flattenItemTrees(rootItems) {
     return item;
   });
 }
+
+export function buildItemTrees(items) {
+  const root = { id: 0, depth: -1, children: [] };
+  const parents = [root];
+
+  _.forEach(items, (item, index) => {
+    const parent = (() => {
+      while (_.last(parents).depth >= item.depth) {
+        parents.pop();
+      }
+      return _.last(parents);
+    })();
+
+    const itemHasChildren = { ...item, children: [] };
+    parent.children.push(itemHasChildren);
+
+    if (item === _.last(items)) {
+      return;
+    }
+
+    if (items[index + 1].depth > item.depth) {
+      delete itemHasChildren.href;
+      parents.push(itemHasChildren);
+    }
+  });
+
+  return root.children;
+}
